+++ registry/src/lib.rs
@@ -4,7 +4,6 @@ use sqlx::{ConnectOptions, Pool, Postgres};
 pub mod config;
 pub mod download;
 pub mod errors;
-pub mod publish;
 pub mod routes;
 pub mod uploader;
 pub mod views;
