+++ registry/src/main.rs
@@ -1,8 +1,8 @@
 use actix_cors::Cors;
 use actix_web::{web, App, HttpServer};
 use trunk_registry::connect;
-use trunk_registry::{config, download, publish, routes};
-use trunk_registry::routes::new_token;
+use trunk_registry::{config, routes};
+use trunk_registry::routes::token::new_token;
 use clerk_rs::{ClerkConfiguration, validators::actix::ClerkMiddleware};
 
 #[actix_web::main]
@@ -30,10 +30,10 @@ async fn main() -> std::io::Result<()> {
             .app_data(web::Data::new(conn.clone()))
             .app_data(web::Data::new(cfg.clone()))
             .app_data(web::Data::new(aws_config.clone()))
-            .service(routes::running)
-            .service(routes::get_all_extensions)
-            .service(publish::publish)
-            .service(download::download)
+            .service(routes::root::ok)
+            .service(routes::extensions::get_all_extensions)
+            .service(routes::extensions::publish)
+            .service(routes::download::download)
             .service(
                 web::scope("/token")
                     .wrap(ClerkMiddleware::new(clerk_cfg))
