+++ registry/src/routes/root.rs
@@ -1,0 +1,6 @@
+use actix_web::{get, HttpResponse, Responder};
+
+#[get("")]
+pub async fn ok() -> impl Responder {
+    HttpResponse::Ok().json("ok")
+}
